moved cached results iter to its own file
