+++ sql/plan/cached_results.go
@@ -18,27 +18,25 @@ import (
 	"errors"
 	"io"
 	"sync"
-	"sync/atomic"
 
 	"github.com/dolthub/go-mysql-server/sql"
 )
 
+// TODO: move this to rowexec package
 var ErrEmptyCachedResult = errors.New("CachedResult contains no rows")
-var ErrRowIterDisposed = errors.New("attempted to call RowIter() on a disposed Node")
 
 // NewCachedResults returns a cached results plan Node, which will use a
 // RowCache to cache results generated by Child.RowIter() and return those
 // results for future calls to RowIter. This node is only safe to use if the
 // Child is deterministic and is not dependent on the |row| parameter in the
 // call to RowIter.
-func NewCachedResults(n sql.Node, mgr *CachedResultsManager) *CachedResults {
+func NewCachedResults(n sql.Node) *CachedResults {
 	return &CachedResults{
 		UnaryNode: UnaryNode{n},
-		Id:        mgr.AllocateUniqueId(),
-		Manager:   mgr,
 	}
 }
 
+// TODO: rewrite this
 // CachedResults tees the child node iterator into an in-memory cache
 // for faster subsequent retrieval. This is usually combined with a
 // HashLookup, whose RowIter defers to a CachedResult child to populate
@@ -57,27 +55,19 @@ func NewCachedResults(n sql.Node, mgr *CachedResultsManager) *CachedResults {
 // fall back to a passthrough iterator.
 type CachedResults struct {
 	UnaryNode
-	Id            uint64
-	Manager       *CachedResultsManager
 	Mutex         sync.Mutex
 	cachedResults []sql.Row
+	// TODO: rewrite or remove
 	//NoCache is set when the memory manager is unable to build
-	// a cache, so we fallback to a passthrough RowIter
+	// a cache, so we fall back to a passthrough RowIter
 	NoCache bool
-	// Finalized is set when we exhaust the child iter, and subsequent
-	// RowIters will read from the cache rather than the child
+	// Finalized is set when the child iter has been exhauster so subsequent RowIters will be read from the cached
+	// result rather than the child iter
 	Finalized bool
-	// Disposed is set after this CachedResults is invalidated
-	Disposed bool
 }
 
 var _ sql.Node = (*CachedResults)(nil)
 var _ sql.CollationCoercible = (*CachedResults)(nil)
-
-func (n *CachedResults) Dispose(ctx *sql.Context) {
-	n.Disposed = true
-	n.Manager.DisposeCachedResultsById(n.Id)
-}
 
 func (n *CachedResults) String() string {
 	pr := sql.NewTreePrinter()
@@ -120,6 +110,7 @@ func (n *CachedResults) IsReadOnly() bool {
 	return n.Child.IsReadOnly()
 }
 
+// TODO: move to rowexec package
 var EmptyIter = &emptyCacheIter{}
 
 func IsEmptyIter(i sql.RowIter) bool {
@@ -133,71 +124,3 @@ var _ sql.RowIter = (*emptyCacheIter)(nil)
 func (i *emptyCacheIter) Next(ctx *sql.Context) (sql.Row, error) { return nil, io.EOF }
 
 func (i *emptyCacheIter) Close(ctx *sql.Context) error { return nil }
-
-// cachedResultsManager manages the saved results collected by CachedResults nodes. It is necessary to do this outside
-// of the CachedResult node instances themselves, since executing a query plan can make transient transforms that are
-// not persisted back and can cause cache memory leaks.
-type CachedResultsManager struct {
-	// cachedResultsCaches tracks caches used by CachedResults globally so that even if a CachedResult
-	// object is copied as part of a transform, its cache can be properly disposed. This is necessary because
-	// when evaluating subquery expressions, the query plan is transformed into a new copy that has prependRow nodes
-	// in it, but that modified version of the query plan is transient, so the caches need to be held somewhere
-	// where they can be properly disposed after the lifetime of an individual subquery expression.
-	cachedResultsCaches map[uint64]*cacheDisposeTuple
-
-	// mutex protects the cachedResultsCaches map from concurrent access.
-	mutex sync.Mutex
-
-	// cachedResultsUniqueIdCounter stores a counter that should only be incremented atomically and is used
-	// as a unique ID when a new CachedResults object is created for a node.
-	cachedResultsUniqueIdCounter uint64
-}
-
-// cacheDisposeTuple is a container for a cache and the related function to dispose it.
-type cacheDisposeTuple struct {
-	cache   sql.RowsCache
-	dispose sql.DisposeFunc
-}
-
-func NewCachedResultsManager() *CachedResultsManager {
-	return &CachedResultsManager{
-		cachedResultsCaches: make(map[uint64]*cacheDisposeTuple),
-	}
-}
-
-func (crm *CachedResultsManager) AllocateUniqueId() uint64 {
-	return atomic.AddUint64(&(crm.cachedResultsUniqueIdCounter), 1)
-}
-
-func (crm *CachedResultsManager) GetCachedResultsById(id uint64) []sql.Row {
-	crm.mutex.Lock()
-	defer crm.mutex.Unlock()
-
-	if results, ok := crm.cachedResultsCaches[id]; ok {
-		return results.cache.Get()
-	} else {
-		return nil
-	}
-}
-
-func (crm *CachedResultsManager) AddNewCache(id uint64, cache sql.RowsCache, dispose sql.DisposeFunc) bool {
-	crm.mutex.Lock()
-	defer crm.mutex.Unlock()
-
-	if _, ok := crm.cachedResultsCaches[id]; ok {
-		return false
-	}
-
-	crm.cachedResultsCaches[id] = &cacheDisposeTuple{cache: cache, dispose: dispose}
-	return true
-}
-
-func (crm *CachedResultsManager) DisposeCachedResultsById(id uint64) {
-	crm.mutex.Lock()
-	defer crm.mutex.Unlock()
-
-	if results, ok := crm.cachedResultsCaches[id]; ok {
-		results.dispose()
-		delete(crm.cachedResultsCaches, id)
-	}
-}
